+++ src/github/copilot_sdk/instrument.clj
@@ -175,6 +175,12 @@
                                 :session-id ::specs/session-id))
   :ret nil?)
 
+(s/fdef github.copilot-sdk.session/disconnect!
+  :args (s/alt :handle (s/cat :session ::specs/session)
+               :explicit (s/cat :client ::specs/client
+                                :session-id ::specs/session-id))
+  :ret nil?)
+
 (s/fdef github.copilot-sdk.session/session-id
   :args (s/cat :session ::specs/session)
   :ret ::specs/session-id)
@@ -276,6 +282,7 @@
                       github.copilot-sdk.session/abort!
                       github.copilot-sdk.session/get-messages
                       github.copilot-sdk.session/destroy!
+                      github.copilot-sdk.session/disconnect!
                       github.copilot-sdk.session/session-id
                       github.copilot-sdk.session/workspace-path
                       github.copilot-sdk.session/get-current-model
@@ -324,6 +331,7 @@
                       github.copilot-sdk.session/abort!
                       github.copilot-sdk.session/get-messages
                       github.copilot-sdk.session/destroy!
+                      github.copilot-sdk.session/disconnect!
                       github.copilot-sdk.session/session-id
                       github.copilot-sdk.session/workspace-path
                       github.copilot-sdk.session/get-current-model